+++ src/test/java/org/folio/circulation/api/support/builders/JsonRequestBuilder.java
@@ -1,0 +1,11 @@
+package org.folio.circulation.api.support.builders;
+
+import io.vertx.core.json.JsonObject;
+
+public class JsonRequestBuilder {
+  protected void put(JsonObject request, String property, String value) {
+    if(value != null) {
+      request.put(property, value);
+    }
+  }
+}