+++ src/test/java/org/folio/circulation/api/support/fixtures/InstanceRequestExamples.java
@@ -3,27 +3,32 @@ package org.folio.circulation.api.support.fixtures;
 import org.folio.circulation.api.support.builders.InstanceRequestBuilder;
 
 public class InstanceRequestExamples {
-  public static InstanceRequestBuilder smallAngryPlanet() {
+  public static InstanceRequestBuilder basedUponSmallAngryPlanet() {
     return create("The Long Way to a Small, Angry Planet");
   }
 
-  public static InstanceRequestBuilder nod() {
+  public static InstanceRequestBuilder basedUponNod() {
     return create("Nod");
   }
 
-  public static InstanceRequestBuilder uprooted() {
+  public static InstanceRequestBuilder basedUponUprooted() {
     return create("Uprooted");
   }
 
-  public static InstanceRequestBuilder temeraire() {
+  public static InstanceRequestBuilder basedUponTemeraire() {
     return create("Temeraire");
   }
 
-  public static InstanceRequestBuilder interestingTimes() {
+  public static InstanceRequestBuilder basedUponInterestingTimes() {
     return create("Interesting Times");
+  }
+
+  public static InstanceRequestBuilder basedUponDunkirk() {
+    return new InstanceRequestBuilder("Dunkirk");
   }
 
   private static InstanceRequestBuilder create(String title) {
     return new InstanceRequestBuilder(title);
   }
+
 }